# changgou-parent/code-template-master/src/main/java/com/changgou/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "微服务工程代码生成器,生成Dao、Service、Controller、Pojo、Feign、Swagger的JSON脚本",
    "version": "1.0.0",
    "title": "深圳黑马训练营代码生成器",
    "termsOfService": "https://github.com/shenkunlin/code-template.git",
    "contact": {
      "email": "shenkunlin@itcast.cn"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "petstore.swagger.io",
  "basePath": "/v2",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/album/{id}": {
      "delete": {
        "tags": [
          "AlbumController"
        ],
        "summary": "根据ID删除Album",
        "description": "根据ID删除Album方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除Album方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除Album"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "AlbumController"
        ],
        "summary": "根据ID修改Album",
        "description": "根据ID修改Album方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Album方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "Album",
            "description": "传入Album的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Album"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改Album"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "AlbumController"
        ],
        "summary": "根据ID查询Album",
        "description": "根据ID查询Album方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Album方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改Album"
            ,
            "schema": {
              "$ref": "#/definitions/Result«Album»"
            }
          }
        }
      }
    },
    "/album/search": {
      "post": {
        "tags": [
          "AlbumController"
        ],
        "summary": "不带分页条件搜索Album",
        "description": "不带分页条件搜索Album方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Album",
            "description": "传入Album的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Album"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索Album"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Album»»"
            }
          }
        }
      }
    },
    "/album/search/{page}/{size}": {
      "get": {
        "tags": [
          "AlbumController"
        ],
        "summary": "分页列表查询Album",
        "description": "分页列表查询Album方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索Album"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Album»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "AlbumController"
        ],
        "summary": "分页条件搜索Album",
        "description": "分页条件搜索Album方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "Album",
            "description": "传入Album的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Album"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索Album"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Album»»"
            }
          }
        }
      }
    },
...
    "/brand/{id}": {
      "delete": {
        "tags": [
          "BrandController"
        ],
        "summary": "根据ID删除Brand",
        "description": "根据ID删除Brand方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除Brand方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除Brand"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "BrandController"
        ],
        "summary": "根据ID修改Brand",
        "description": "根据ID修改Brand方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Brand方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "Brand",
            "description": "传入Brand的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Brand"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改Brand"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "BrandController"
        ],
        "summary": "根据ID查询Brand",
        "description": "根据ID查询Brand方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Brand方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改Brand"
            ,
            "schema": {
              "$ref": "#/definitions/Result«Brand»"
            }
          }
        }
      }
    },
    "/brand/search": {
      "post": {
        "tags": [
          "BrandController"
        ],
        "summary": "不带分页条件搜索Brand",
        "description": "不带分页条件搜索Brand方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Brand",
            "description": "传入Brand的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Brand"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索Brand"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Brand»»"
            }
          }
        }
      }
    },
    "/brand/search/{page}/{size}": {
      "get": {
        "tags": [
          "BrandController"
        ],
        "summary": "分页列表查询Brand",
        "description": "分页列表查询Brand方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索Brand"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Brand»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "BrandController"
        ],
        "summary": "分页条件搜索Brand",
        "description": "分页条件搜索Brand方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "Brand",
            "description": "传入Brand的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Brand"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索Brand"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Brand»»"
            }
          }
        }
      }
    },
    "/category": {
      "get": {
        "tags": [
          "CategoryController"
        ],
        "summary": "查询所有Category",
        "description": "查询所有Category方法详情",
        "operationId": "findAllUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "查询所有Category"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Category»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "CategoryController"
        ],
        "summary": "添加Category",
        "description": "添加Category方法详情",
        "operationId": "addUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "添加Category方法详情",
            "required": true
              ,
              "schema": {
                "$ref": "#/definitions/Category"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "添加Category"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      }
    },
    "/category/{id}": {
      "delete": {
        "tags": [
          "CategoryController"
        ],
        "summary": "根据ID删除Category",
        "description": "根据ID删除Category方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除Category方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除Category"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "CategoryController"
        ],
        "summary": "根据ID修改Category",
        "description": "根据ID修改Category方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Category方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "Category",
            "description": "传入Category的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Category"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改Category"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "CategoryController"
        ],
        "summary": "根据ID查询Category",
        "description": "根据ID查询Category方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Category方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改Category"
            ,
            "schema": {
              "$ref": "#/definitions/Result«Category»"
            }
          }
        }
      }
    },
    "/category/search": {
      "post": {
        "tags": [
          "CategoryController"
        ],
        "summary": "不带分页条件搜索Category",
        "description": "不带分页条件搜索Category方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Category",
            "description": "传入Category的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Category"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索Category"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Category»»"
            }
          }
        }
      }
    },
    "/category/search/{page}/{size}": {
      "get": {
        "tags": [
          "CategoryController"
        ],
        "summary": "分页列表查询Category",
        "description": "分页列表查询Category方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索Category"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Category»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "CategoryController"
        ],
        "summary": "分页条件搜索Category",
        "description": "分页条件搜索Category方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "Category",
            "description": "传入Category的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Category"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索Category"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Category»»"
            }
          }
        }
      }
    },
...
    "/categoryBrand/search": {
      "post": {
        "tags": [
          "CategoryBrandController"
        ],
        "summary": "不带分页条件搜索CategoryBrand",
        "description": "不带分页条件搜索CategoryBrand方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "CategoryBrand",
            "description": "传入CategoryBrand的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/CategoryBrand"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索CategoryBrand"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«CategoryBrand»»"
            }
          }
        }
      }
    },
...
    "/para/search": {
      "post": {
        "tags": [
          "ParaController"
        ],
        "summary": "不带分页条件搜索Para",
        "description": "不带分页条件搜索Para方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Para",
            "description": "传入Para的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Para"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索Para"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Para»»"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Result": {
      "type": "object",
      "required": [
        "code",
        "data",
        "flag",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,20001:失败,20002:用户名或密码错误,20003:权限不足,20004:远程调用失败,20005:重复操作,20006:没有对应的数据"
        },
        "data": {
          "type": "object",
          "description": "逻辑数据"
        },
        "flag": {
          "type": "boolean",
          "example": false,
          "description": "执行是否成功,true:成功,false:失败"
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "Result"
    },
...
    "Result«Album»": {
      "type": "object",
      "required": [
        "code",
        "data",
        "flag",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,20001:失败,20002:用户名或密码错误,20003:权限不足,20004:远程调用失败,20005:重复操作,20006:没有对应的数据"
        },
        "data": {
          "description": "逻辑数据",
          "$ref": "#/definitions/Album"
        },
        "flag": {
          "type": "boolean",
          "example": false,
          "description": "执行是否成功,true:成功,false:失败"
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "Result"
    },
    "Album": {
      "type": "object",
      "properties": {
        "id": {
          "type": "long",
          "description": "编号"
        },
        "title": {
          "type": "string",
          "description": "相册名称"
        },
        "image": {
          "type": "string",
          "description": "相册封面"
        },
        "imageItems": {
          "type": "string",
          "description": "图片列表"
        }
      },
      "description": "Album"
    },
...
  },
...
}